Add support for event listeners
Previously, the `Listen` method was not implemented, and it simply
returned `nil`. Consequently, any uses of the method will block
further code execution indefinitely.

This commit adds basic support for event listeners.
It creates an event listeners registry / pool on the target handler
which `Listen` registers against. Within the `processEvent` method,
it will loop through all the registered event listeners, and call
them if the event matches. The event listener channel will be given
the event payload.

It can be used for creating a custom action that waits for the
page to load, for example:

```
func WaitLoadEvent() cdp.Action {
    return cdp.ActionFunc(func(ctx context.Context, handler cdptypes.Handler) error {
        select {
        case <-handler.Listen(cdptypes.EventPageLoadEventFired):

        case <-ctx.Done():
            return ctx.Err()
        }
        return nil
    })
}
```

The type can be determined using `.(type)`.

The above assumes the following import:

```
cdp "github.com/knq/chromedp"
```