Fix empty struct for CDP object types with no properties
`Headers` under the `network` package is an empty struct.

This is because the generator did not account for cases where
an object definition contains no properties.

In these cases, we can simply use `map[string]interface{}` so
that the type is treated as a generic key-value object.

This commit should allow extra HTTP headers to be set as
this was not possible with an empty type struct.

Closes #68