Process events synchronously (#75)
Proof of concept for processing events synchronously.

Two changes:

* Do not run updates and event handlers each in their own goroutine, but
  instead run them in one goroutine in the order they came in. This
  avoids the race described in #75.
* Split the process queues into their own goroutines. This is necessary
  so that event handling can run commands. If all the queues are
  processed from the same goroutine, then when for exapmle when
  `documentUpdated()` runs,  it tries to run a `dom.GetDocument()`,
  which would deadlock because documentUpdated() is occupying the
  handler goroutine.

This patch does not represent a complete fix, because I believe it may
make a lot of locking and other concurrent machinary unnecessary. I
don't have time to explore that right now, but I don't think it's
harmful as far as I can tell.

Ignore EventFrame* state changes
--------------------------------

These state changes are intended to be written to a `*cdp.Frame` which
only arrives when `EventFrameNavigated` is received. This latter event
may never occur, or happen after the state change is written.

I'm not convinced they are very useful, but their precense causes a
deadlock if handling events synchronously, because it waits for a frame
that never comes.

Please see discussion in #75 and comments in the code to understand the detail here.

Use a large buffer size for qevents
-----------------------------------

See comment within the code, and discussion in #75.