Ignore EventFrame* state changes
These state changes are intended to be written to a `*cdp.Frame` which
only arrives when `EventFrameNavigated` is received. This latter event
may never occur, or happen after the state change is written.

I'm not convinced they are very useful, but their precense causes a
deadlock if handling events synchronously, because it waits for a frame
that never comes.

Please see discussion in #75 and comments in the code to understand the detail here.