Detect support for --headless and use if available
Chrome 59 gained support for `--headless`.

This PR makes the runner use this if `Headless` is supplied to the
runner, if it's available, falling back to the old method
(`headless_shell`) if it is unavailable.

Unfortunately it does not seem possible to get chrome to list which
command line options it supports. I could not find a way of detecting
the feature, so we instead just use a check on the version number, which
is not the preferred solution but does the job.