add the first simple benchmark
This parallel benchmark simply opens a new tab on a browser, navigates
to a local page, and waits for an element to be visible. Will be useful
to monitor chromedp's overhead, for example when it comes to marshaling
and unmarshaling json objects when talking to Chrome.

name           time/op
TabNavigate-8  30.9ms ± 3%

name           alloc/op
TabNavigate-8   458kB ± 0%

name           allocs/op
TabNavigate-8   1.19k ± 0%