try to fix the TestBrowserQuit flake on Travis
It would sometimes fail with the "did not expect a nil error" failure.
Intuitively, this could happen if Chrome takes too long to respond to
os.Interrupt, and actually succeeds at loading the page.

To make that impossible, send it os.Kill, which should exit the process
immediately. If the flakes still happen, we can consider adding an extra
wait on the process, or some form of sleep.

Another bonus is that os.Kill works on Windows, so the test is now
portable.