remove detached sessions from Browser
Before this, it was possible to create and cancel targets on a browser,
and the map[target.SessionID]*Target would keep growing without ever
shrinking back down.

This wasn't caught by any of the tests or benchmarks though, as the map
was initialized with a size of 1024 and the benchmark was usually run
with N=150.

In any case, this implementation is more correct, doesn't leak memory,
and will scale better. No noticeable benchmark changes.