have RemoteAllocator cancel on a dropped ws connection
ExecAllocator did this already, but not RemoteAllocator. It was a bit
tricky, as its Allocate happens with a completely separate context, so
we can delay closing the websocket on cancel until the pages have been
closed.

To work around the issue, signal that the browser connection has been
lost via a channel field on Browser, and not via Context. This works
even when the context used in Allocate doesn't contain anything related
to chromedp.

This also means that Allocator implementations can have the same feature
from third party packages, as the LostConnection field is exported.