don't break pages which load child frames quickly
If a page loads a child frame before the top frame has finished loading,
we'd treat the child frame as the top-level frame, and the original
frame would never get all of its nodes.

This can be seen in the addition to TestLoadIframe; before the fix, that
test would hang.

The problem was that we treated all navigated frames as top-level
frames, when that's not true. Only when a newly navigated frame has no
parent frame should we treat it as our new top-level frame.

Fixes #304.