don't panic when a cancel happens mid-Allocate
NewContext was racy with ExecAllocator.Allocate.

First, Allocate incremented Context.wg, which is a racy thing to do
since context cancellation might be happening concurrently. To fix that,
separate the wait for Browser into a Mutex, which allows for this
behavior.

Second, we might call cmd.Wait before cmd.Start, which is a data race on
one of cmd's fields. Ensure that cmd.Start was called before we start
the cancellation goroutine.

Finally, add a test that spotted this issue with -race, and is now
happy.