make ExecAllocator use --no-sandbox when root
Chrome will refuse to run as the root user, unless the --no-sandbox flag
is given:

	Running as root without --no-sandbox is not supported.
	See https://crbug.com/638180.

If the user didn't specify the flag in any way, and we're root, add the
flag automatically.

Also add a regression test with 'docker run', to really test that
ExecAllocator can properly allocate a Chrome running as root.

Fixes #297.