support running actions in listener callbacks
This is necessary to set up a listener function to automatically accept
or reject opened dialogs, since otherwise the target handler would get
stuck on a wait function.

We still want event callbacks to be synchronous, but they don't have to
be on the same goroutine that updates the handler and runs wait
functions. Separating that goroutine makes the listener callbacks more
powerful.

Add the first dialog test, courtesy of Randy Cahyana.