remove "eventQueue is full" panic
Added a test case that reproduced the panic fairly reliably. This could
happen realistically if a machine is under stress, and a target receives
a large number of events before they can be handled.

It's fine for the browser to get blocked sending to page.eventQueue;
after all, this will give the relevant Target to catch up, and the
browser will be unblocked at a later time.

Fixes #322.