deflake TestTargets and TestRemoteAllocator
The first was due to a bug in how we used the initial blank tab from a
newly allocated browser. For now, fix the race with a bit of polling,
and leave a TODO for a better future fix.

The second was simply that the test was badly written, falling victim to
a similar race. It can be rewritten to be stable.