support using tabs opened by existing tabs
For the initial support, all we need is to receive an event when such a
new tab is opened, and to be able to attach to said new tab (target).

Do precisely these two things, adding the WithTargetID API. Add a simple
test too, which is probably a good base for a future example.

Fixes #317.