rewrite Screenshot to work on all scale factors
The old Screenshot action used a bit of JavaScript and arithmetic to
properly capture the screenshot of a ndoe.

This worked fine for simple scenarios with a scaling factor of 1.0, but
broke horribly in other scaling factors, such as 1.5 or 2.0 when on
high-DPI devices.

Chrome added a "clip" parameter in its screenshot command in mid-2017 to
be able to take a screenshot of a specific viewport, so do exactly that.
This lets us write simpler code that works on all DPIs. Add a test that
ensures that.

As a bonus, this allows dropping the external "imaging" dependency, used
to crop the larger PNG before.

Fixes #191.