remove a goroutine from Browser.run
Turns out we don't need an extra goroutine to handle tabs. The idea here
was that we could handle tab management and tab event routing
separately, but in practice that doesn't measurably improve performance.

If anything, adding another goroutine makes the code more confusing to
follow, and adds a bit of overhead.