clarify tab queue, deflake TestListenBrowser
The tab queue takes messages, not event data, so the new name is
clearer.

TestListenBrowser isn't racy now, but I have found it to be racy when
trying out some changes. The browser is shared, so we can't be sure that
a regular int counter is safe.

Finally, remove an obsolete TODO.