use listeners for all command responses
Adding command response messages to the Listen API has one big
advantage; we can drop all the respByID code, and instead have both
Execute functions set up a listener to wait for the right response.

This also greatly simplifies cmdQueue, as it can now simply contain the
message to write over the transport.

This change does make the Execute functions slightly more expensive, as
they need to set up a context and a listener. However, we don't expect
that users will run thousands of actions at once on a single target, so
that's a non-issue.

Finally, this change requires listeners to be run asynchronously, as
otherwise the target handler can deadlock itself. That's an OK change to
make, as the documentation makes no promises about how Listen callbacks
work.