revert Listen callbacks to being synchronous
Using a goroutine is actually racy, as two callbacks on a single
listener might happen at once.

We only did this so that we could support running actions without an
extra goroutine in the Listen func, just like before. But there's no
value in keeping that property, particularly since we haven't tagged a
release with that API yet.