increase timeout when waiting for new page targets
We'll remove this code eventually, but for now, 50ms is turning out to
not be nearly enough when the machine is slow or under load.

Double each of the sleeps and multiply the retries by ten, meaning that
we wait for up to a full second. This should be enough for practically
all scenarios for now.

Fixes #352.