rewrite forceIP to not be string-based
Using the net and net/url packages is more robust. Perhaps it adds a
couple of allocations, but this isn't code that will run often, and
having it behave correctly is more important.

For example, I was seeing errors before, as it didn't handle IPv6
hostnames. That's fixed now, though it didn't affect any tests.