stop running wait funcs synchronously
This worked for most programs, but had one major pitfall. If a large
query was run on a large page, this could trigger thousands of DOM
events, filling up all internal buffered channels. This would in turn
block the query forever, as it would never see its response message.

Add a test to reproduce the hang, and fix it. For now, this means adding
a mutex. We might get rid of it at some point.

Fixes #375.