revert back to using one goroutine per selector
The point of using a single ticker per browser and a func channel was so
that the wait funcs could run synchronously with the target handler. On
paper, this had several advantages:

* Less use of mutexes
* Fewer goroutines
* Better scheduling, as the target handler sees the incoming events

However, none of those turned out to be true in practice. For example,
less mutexes and goroutines don't matter, because in practice there are
usually only one or two selectors per target at a time.

More importantly, this synchronous design was more complex and
expensive. It requires a ticker and one channel per target, which are
used even if there are no running selectors at the moment.

This could mean that the CPU would be spinning a bit, which is more
noticeable on small computers.

While at it, move the goroutine outside of Selector.run, so that the
code doesn't need the extra level of indentation.

Updates #380.