also capture targetInfoChanged for a browser's first tab
We were seeing flakes when allocating new browsers, particularly with
headless-shell. We were never seeing the target ID for the blank page
that a browser opens up with.

This was because we were only capturing targetCreated events. For
example, if the page is created first, and then its info is changed to
"about:blank", we wouldn't catch it.

What we need is an almost exact copy of the code we have in
WaitNewTarget. Not exactly the same code, though, so a bit of copy
pasting is fine.

There isn't a regression test for this one, as I haven't been able to
reproduce it reliably with regular chromium. However, headless-shell
reproduced it via 'go test' most of the time, and CI already caught
that.