rethink the device emulation API a bit
First, we don't need named func types for all the options. A type alias
is enough for our needs, and makes everything much simpler for the
users.

For example, in emulate.go and device/device.go we had to use explicit
types and copy slices, because the device package can't depend on the
parent package's named type. However, if the type is an alias, the
device package can declare the alias too, fixing both issues.

Second, split ViewportParams into UserAgent and Viewport. The old method
returned parameters for two actions, even though it was documented as
only returning parameters for EmulateViewport. Splitting UserAgent makes
sense in this context, and keeps the signature simpler and shorter.

Rename the remaining method to Viewport too, as the "Params" bit is a
bit redundant.

Third, make the zero value Device be invalid, to avoid people using
BlackberryPlayBook by default via "var dev device.Device" and such. This
also makes the generated code a bit nicer, as the first real device line
isn't any different from the others. Add a test, too.

Finally, both Emulate and EmulateViewport can return a complex action
via Tasks; there's no need for the extra complexity of an ActionFunc.