don't make ExampleEmulate a runnable example
It requires an internet connection, so it was making 'go test' either
painfully slow or just fail while I was working on a spotty connection.

If we want it as a runnable example in this repo, it should be rewritten
to visit a local html file instead. Otherwise, it can stay as-is, or be
moved to chromedp/examples.