target: fix issue with DOM child inserted events
A small issue with the logic of the insertNode nodeOp, and present since
the beginning of the project, meant that element nodes that were copied
from the browser to the Target's Frames/Nodes, that later had a child
element added at the _front_ of the element's children would instead
have the child element node added to the _end_ of the list.

This resulted in incorrect generation of the Node's FullXPath(), and
could cause frustrating/hard to track errors when later attempting to
access that Node using the element query actions.

This change fixes the single one line logic, and adds an
Example_DocumentDump that works as a test for this specific case.
Additionally updates to the latest cdproto in order to make use of
Node.Dump in the example test case.

Note: the Example_DocumentDump test may be flaky at the moment, as it
includes the NodeIDs of the dumped DOM elements, which can vary from
Chrome runs. This is planned to be fixed before the next release.

Fixes #412.