replace all coordinate APIs to use float64
As of Chromium 76.0.3809.87, some click tests would start failing:

	--- FAIL: TestMouseClickOffscreenNode/02 (0.23s)
	    input_test.go:157: got error: json: cannot unmarshal number 490.28570556640625 into Go value of type int
	--- FAIL: TestMouseClickOffscreenNode/03 (0.19s)
	    input_test.go:157: got error: json: cannot unmarshal number 490.28570556640625 into Go value of type int
	--- FAIL: TestMouseClickOffscreenNode/01 (0.21s)
	    input_test.go:157: got error: json: cannot unmarshal number 490.28570556640625 into Go value of type int

The problem here was unmarshalling positions into []int; use []float64.

To be consistent with all the other coordinate APIs, change all the
input APIs to also use float64 instead of int64. This reflects the
actual floating-point precision that Chrome supports internally, and
allows chromedp users to better support high-dpi displays.

This is a breaking change for these two APIs, but we're still at a v0,
and the fix for users should be trivial.