don't panic when reusing actions in many Run calls
This is perhaps not common, but still valid use of chromedp.

In one of the refactors in the past few months, we rewrote waitReady to
use a channel. Allocating the channel outside the function closure was
an attempt at reusing the allocation, but it was wrong.

In particular, it could lead to the same channel being used again if the
action was re-run, meaning a panic as we'd try to close that same
channel twice.

For now, simply allocate it inside the closure, and add a regression
test.

Fixes #452.