fix TestScreenshotHighDPI after some recent failures
The element we were taking a screenshot of was no longer in view, for
some reason. I suspect because of some recent change in Chrome that
means the page is rendered differently.

In any case, it doesn't matter. The size we were using was arbitrary, so
just make it a bit larger.