make ExecAllocator more robust with bufio.Reader
We were reading the browser's output lines to find the websocket URL
with plain reads. This is fine, but a bit tricky; if we happen to get a
line split in two reads for whatever reason, we could end up with a
broken websocket URL.

Instead, use bufio.Reader, which supports reading entire lines at once.

This makes the code more robust, but also a tiny bit simpler.