stop eating exception events in the browser
This was initially done in 7c1a9fbf3e for the purpose of not hiding
exceptions in our tests. However, we went way overboard; we can expose
them in our tests, without exposing them for everyone.

Worse even, we weren't letting the events through, so ListenBrowser and
ListenTarget never received exceptions.

Add this to ExampleListenTarget_consoleLog, so that users can see how to
respond to exceptions, and to act as a test.

Fixes #519.