add support for graceful browser closing via Cancel
The added test uses cookies; without graceful browser closing, the test
failed most of the time due to a race between the cookie being saved and
the browser being killed. It now always succeeds.

Big thanks to Paul Murley for the suggestion, and to alanrothschild for
providing a way to write a reliable test.

Fixes #472.