fix a channel race in Cancel's graceful shutdown
In c661502af3 we added support for graceful browser shutdown via Cancel;
instead of cancelling the context directly and killing the browser
process, we instead send a browser.CommandClose command. This way, if
the browser had to store any of its state to disk, or was in the middle
of any sensitive operation, no data will be lost or corrupted.

However, there was a channel race, as spotted numerous times by CI:

	    TestGracefulBrowserShutdown: chromedp_test.go:717: want cookies "cookie1=value1"; got ""
	--- FAIL: TestGracefulBrowserShutdown (0.55s)

The timeline was as follows:

1) Cancel is called
2) CommandClose is sent to the browser
3) Cancel blocks, waiting for the browser process to stop
4) The browser process begins shutdown, and closes the websocket connection
5) The Browser.run goroutine notices the lost connection, and signals LostConnection
6) ExecAllocator.Allocate's LostConnection goroutine sees the signal,
   and immediately cancels the browser context, killing the process
7) Cancel gets unblocked and returns

The problem here was that the goroutine in step 6 thinks that we just
suddenly lost our websocket connection. In that scenario, it's
reasonable to cancel the entire context; all we can do is clean up the
resources and error.

However, that's not the case here, since we're doing a graceful
shutdown. Instead, add another channel which is used to signal that
we're in the middle of a graceful shutdown. And if so, the goroutine in
step 6 will not do an immediate cancel.

Before this fix, this bug manifested itself about 10% of the time on my
laptop with headless-shell under Docker via TestGracefulBrowserShutdown.
After the fix, I can easily get to 100 runs with no failures. The
problem presumably manifested itself more frequently on CI due to the
machine being under more stress, or having fewer resources available.

For #472.