Enable filtering visible nodes as an option
Currently, we cannot tell ChromeDP to select all visible nodes without
waiting for them.

If one calls
```
NodeIDs(selector, WaitVisible, AtLeast(0))
```

chromeDP selects all nodes with the selector and then wait for them to be visible.
This might hang if it never happens and the node keeps being hidden.

With this PR, we can do
```
NodeIDs(selector, FilterVisible, AtLeast(0))
```

However, the following will still hang, since the waiting is done before the filtering;
as we need the nodes to be at least ready to filter them. I think it's OK,
but it might be confusing.

```
NodeIDs(selector, WaitVisible, FilterVisible, AtLeast(0))
```