initial support for iframe queries via FromNode
If a query uses the FromNode option, and the provided node is an iframe,
properly use its content document as the node to run the query on.

With this change, TestQueryIframe now succeeds, and existing behavior is
kept intact. Note that we don't introduce any new API; we simply reuse
the existing FromNode query option.

The API is still a bit clunky, since we require a first call to Nodes
via Run to pick up the iframe node. We might consider a bit of
higher-level API to make that simpler.

Updates #72.