introduce RunResponse to observe page navigations
Until now, the API only provided a nice way to wait for navigations to
finish if they were directly triggered via Navigate. Doing the same for
a Click on a link, for example, required setting up ListenTarget and
manually waiting for the load event to show up.

Another common need was to check if a navigation ended up in error
status codes such as 404 or 503. Because of how a browser works, those
are still successful navigations, since the site might have a proper 404
landing page.

Fix both of those problems by adding a version of Run which waits for a
page navigation to finish, and returns the HTTP response information for
its document.

The code also handles the case where loading a page fails, suchas if
connecting to the server results in a TLS error. Add tests for such edge
cases, as well as for HTTP redirects, which require no extra code.

For more details, see the RunResponse docs, as well as
ExampleRunResponse.

Fixes #641.