make RunResponse more resilient to iframes and races
This comes with a very beefy new test, which includes many test cases
for both Navigate and Click, and includes an iframe that keeps reloading
in the background.

Before this fix, the added test would fail pretty often, and
TestClickNavigate would fail about 10% of the time. Now they all
reliably pass.

The added code is pretty complex, but well documented. The basic idea is
to use frame lifecycle events to keep track of when a new page load
begins, and then use that loaderID to find the right request/response
and load event.