revert the previous commit, with a regression test
If the parent context was cancelled at the right time - after the
actions finished, but before the load finshed - then the select would
race. In the 50% of cases where we would enter the first case, loadErr
would most likely be nil, in which case we would just "return nil".

This is wrong, because if the parent context cancels, we need to report
that as an error. Returning a nil error makes it seem like the
navigation load succeeded, which is definitely not the case here.

The return is now very well documented, and the test includes docs as
well.