avoid nil pointer dereference panics in enclosingFrame
This should make the code more resilient; the crash was encountered in
some complex sites, but we were unable to reproduce it with a small HTML
test page.

The current best theory is that a node was used after it's replaced, so
its parent is no longer present in the frame's Nodes map. In that case,
return an empty FrameID, and the selector will try again with the new
and correct node.