Fix deadlock in TestPrematureCancel.
The `c.allocated` channel is only written in `NewContext`, and
subsequently closed when the browser is allocated. But the browser is
allocated in `Run` and this test calls `Cancel` before `Run`, so the
channel couldn't possible receive any more data nor be closed for a
second read to not block.