make NewRemoteAllocator accept url without "devtools/browser/..."
A full websocket debugger URL is something like this:
ws://127.0.0.1:9222/devtools/browser/120e6fb7-0094-47a2-8b4c-51741797f280
And the ID part is not static, which requires extra work to get
the correct URL for NewRemoteAllocator.

This change allows the client to just pass the base URL to
NewRemoteAllocator. It will send a request to "/json/version"
to get the full websocket debugger URL.

After the change, the URL with the following formats are accepted:
* ws://127.0.0.1:9222/
* http://127.0.0.1:9222/

But "ws://127.0.0.1:9222/devtools/browser/" are not accepted.
Because it contains "/devtools/browser/" and will be considered
as a valid websocket debugger URL.