dispatch input and change events at the SetValue() target (#819)
chromedp.SetValue is supposed to be called against input, textarea and select.
For these kind of elements, dispatching input and change events after their
values are changed is required.

The document says chromedp.SetValue supports other elements with a .value field.
The input and change events will be fired for those elements too. But I'm not ware
of such an element, so there is not unit test to cover this case.