set chromedp.ByQuery option explicitly in Example_retrieveHTML
It's not intuitive that the default query option is BySearch.
Most users will think that the sel is a css selector (ByQuery or
ByQueryAll). Both BySearch and ByQuery work in the example, but
very likely BySearch does not work in user code if the sel is a
css selector. Since #content is a css selector, it's better to
set ByQuery explicitly.

Regarding ByID and ByQuery, ByID is a special case of ByQuery,
except that it tries to trim # first and then add prefix #.
I think the use of ByID should not be encouraged.