introduce BrowserContext support
A BrowserContext is similar to an incognito profile but you can have
more than one. Don't get confused by the name, it has nothing to do
with the Go package "context".

Two more chromedp.ContextOption's are added:

- WithNewBrowserContext: to create a target in a new BrowserContext.
- WithExistingBrowserContext: to create a target in an existing
  BrowserContext.

If the context is created with the WithNewBrowserContext option, the
id of the new BrowserContext can be retrived by:
  chromedp.FromContext(ctx).BrowserContextID