add RetryInterval option to query actions (#1159)
A query action will retry every 5ms if it fails to get the expected nodes.
The family of Wait* functions depend on this behavior. The problem is,
when it waits too long, there will be many retries. This new option allows
the caller to change the default interval to reduce retries.

That said, the poll action is still preferable if the caller needs to wait.