Fix: Reinstate a step to stop the chromedp-125.0.6422.142 container
Oops. We need to reinstate the explicit step to stop and remove the
chromedp-125.0.6422.142 container.

This is because the chromedp-latest container will listen on the same
port (9222) and we need that port to be free on the host, before
starting the chromedp-latest container.

The "Stop Conaainers" step has been revised so that it now only stops
the chromedp-latest container.